code/robot3/requirements.txt
absl-py==2.4.0
astunparse==1.6.3
audioop-lts==0.2.2
audioread==3.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
colorzero==2.0
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
evdev==1.9.3
filelock==3.25.0
flatbuffers==20181003210633
fonttools==4.61.1
fsspec==2026.2.0
gast==0.7.0
google-pasta==0.2.0
gpiozero==2.0.1
grpcio==1.78.0
h5py==3.15.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
keras==3.13.2
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.46.0
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
namex==0.1.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pillow==12.1.1
platformdirs==4.9.2
polars==1.38.1
polars-runtime-32==1.38.1
pooch==1.9.0
protobuf==7.34.0
psutil==7.2.2
PyAudio==0.2.14
pycparser==3.0
pydub==0.25.1
Pygments==2.19.2
pynput==1.8.1
pyparsing==3.3.2
pyserial==3.5
python-dateutil==2.9.0.post0
python-xlib==0.33
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
RPi.GPIO==0.7.1
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.0
six==1.17.0
sounddevice==0.5.5
soundfile==0.13.1
soxr==1.0.0
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
threadpoolctl==3.6.0
torch==2.10.0
torchvision==0.25.0
typing_extensions==4.15.0
ultralytics==8.4.20
ultralytics-thop==2.0.18
urllib3==2.6.3
Werkzeug==3.1.6
wheel==0.46.3
wrapt==2.1.1
